Update почты и всего, что с ней связано ПРоверила- приложение запускается
