+++ src/main/java/doctorBookingApp/controller/UserController.java
@@ -1,11 +1,12 @@
  package doctorBookingApp.controller;
 
- import doctorBookingApp.dto.NewUserDTO;
  import doctorBookingApp.dto.StandardResponseDto;
- import doctorBookingApp.dto.UserDTO;
+ import doctorBookingApp.dto.usersDTO.NewUserDTO;
+ import doctorBookingApp.dto.usersDTO.UserDTO;
  import doctorBookingApp.exeption.RestException;
- import doctorBookingApp.service.ConfirmationCodeService;
- import doctorBookingApp.service.UserService;
+ import doctorBookingApp.service.userService.ConfirmationCodeService;
+ import doctorBookingApp.service.userService.RegistrationUserService;
+ import doctorBookingApp.service.userService.UserService;
  import io.swagger.v3.oas.annotations.Operation;
  import io.swagger.v3.oas.annotations.media.Content;
  import io.swagger.v3.oas.annotations.media.Schema;
@@ -15,6 +16,8 @@
  import lombok.RequiredArgsConstructor;
  import org.springframework.http.HttpStatus;
  import org.springframework.http.ResponseEntity;
+ import org.springframework.stereotype.Controller;
+ import org.springframework.ui.Model;
  import org.springframework.web.bind.annotation.*;
 
  @RestController
@@ -24,27 +27,74 @@
 
      private final UserService userService;
      private final ConfirmationCodeService confirmationCodeService;
+     private final RegistrationUserService registrationUserService;
 
 
      //РЕГИСТРАЦИЯ ПОЛЬЗОВАТЕЛЯ
+     //Контроллер принимает запрос на подтверждение, вызывает метод подтверждения
+     // и возвращает соответствующий ответ.
 
+
+     @Operation(summary = "Регистрация нового пользователя")
+     @ApiResponses(value = {
+             @ApiResponse(responseCode = "200", description = "Регистрация практически завершена, проверьте свою электронную почту.",
+                     content = @Content(mediaType = "application/json")),
+             @ApiResponse(responseCode = "409", description = "Пользователь с таким email уже существует",
+                     content = @Content(mediaType = "application/json"))
+     })
 
      @PostMapping("/register")
-     public ResponseEntity<String> registerUser(@RequestBody NewUserDTO newUser) throws MessagingException, RestException {
-         userService.registrationUser(newUser);
-         return ResponseEntity.ok("Регистрация практически закончена. Проверьте свой адрес электронной почты на наличие кода подтверждения");
+     public ResponseEntity<String> registerUser(@RequestBody NewUserDTO newUser) throws RestException, MessagingException {
+         registrationUserService.registrationUser(newUser);
+         return ResponseEntity.ok("Регистрация практически завершена. Проверьте свой электронный почтовый ящик на наличие кода подтверждения.");
      }
+
+
+
+     @Operation(summary = "Регистрация. Подтверждение пользователя по конфирмационному коду")
+     @ApiResponses(value = {
+             @ApiResponse(responseCode = "200", description = "Пользователь успешно подтвержден",
+                     content = @Content(mediaType = "application/json")),
+             @ApiResponse(responseCode = "400", description = "Код не найден или срок его действия истек",
+                     content = @Content(mediaType = "application/json"))
+     })
 
 
      @PostMapping("/confirm")
      public ResponseEntity<?> confirmUser(@RequestParam String confirmationCode) {
          try {
-             UserDTO userDTO = confirmationCodeService.confirmation(confirmationCode);
-             return ResponseEntity.ok(userDTO);
+             registrationUserService.confirmUser(confirmationCode);
+             return ResponseEntity.ok("Пользователь успешно подтвержден");
          } catch (RestException e) {
-             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Неверный код подтверждения");
+             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Код не найден или срок его действия истек");
          }
      }
+
+     //@RestController возвращает данные непосредственно в виде JSON/XML, а не перенаправляет на HTML страницу.
+     //Для перенаправления на HTML страницу лучше использовать @Controller.  ???????????? надо ли менять
+
+     @GetMapping("/confirm")
+     public String confirmUserPage(@RequestParam String confirmationCode, Model model) {
+         model.addAttribute("confirmationCode", confirmationCode);
+         return "confirm"; // Имя HTML страницы подтверждения
+     }
+
+
+
+
+//
+//     @PostMapping("/confirm")
+//     public ResponseEntity<?> confirmUser(@RequestParam String confirmationCode) {
+//         try {
+//             UserDTO userDTO = confirmationCodeService.confirmation(confirmationCode);
+//             return ResponseEntity.ok(userDTO);
+//         } catch (RestException e) {
+//             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Код не найден или срок его действия истек");
+//         }
+//     }
+//
+
+
 
 
      //ПОЛУЧЕНИЕ ИНФОРМАЦИИ О ПОЛЬЗОВАТЕЛЕ
