+++ src/main/resources/application.properties
@@ -1,7 +1,8 @@
 spring.application.name=doctor-booking-app
 spring.datasource.url=jdbc:mysql://localhost:3306/db108
-spring.datasource.username=root
-spring.datasource.password=root12345
+spring.datasource.username=Natali
+spring.datasource.password=New-IT_2024
+
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 
@@ -21,7 +22,7 @@ jwt.lifetime=3600000
 
 #Email ?????????
 spring.mail.username=doctorbooking80@gmail.com
-spring.mail.password=Root-12345
+spring.mail.password=nlgzbiimnmxkxbnu
 spring.mail.host=smtp.gmail.com
 spring.mail.port=587
 spring.mail.properties.mail.smtp.auth=true
